pyproject.toml
[tool.poetry]
name = "repl-nix-flight-deals-start"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
twilio = "^9.3.1"
watchdog = "^5.0.2"
amadeus = "^10.1.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
